* Rename `common` to `bootstrap` (meaningful naming)
* Move storage initialisation from `gateways` to `bootstrap`

* Update documentation
