+++ src/alembic/env.py
@@ -1,8 +1,8 @@
 import logging
 from asyncio import get_event_loop
 
-from common.bootstrap import application_init
-from common.config import AppConfig
+from bootstrap.bootstrap import application_init
+from bootstrap.config import AppConfig
 from sqlalchemy.ext.asyncio import AsyncEngine
 
 from alembic import context
