+++ src/bootstrap/__init__.py
@@ -1,0 +1,2 @@
+from .bootstrap import application_init
+from .config import AppConfig
