+++ src/bootstrap/bootstrap.py
@@ -2,11 +2,11 @@ from celery import Celery
 from dependency_injector.containers import DynamicContainer
 from dependency_injector.providers import Object
 from domains import init_celery, init_domains
-from gateways.storage import init_storage
 from pydantic import BaseModel, ConfigDict
 
 from .config import AppConfig, init_logger
 from .di_container import Container
+from .storage import init_storage
 
 
 class InitReference(BaseModel):
