+++ src/bootstrap/di_container.py
@@ -7,9 +7,9 @@ from domains.books._data_access_interfaces import (
 from domains.books.models import BookModel
 from gateways.event import NullEventGateway
 from sqlalchemy_bind_manager import SQLAlchemyBindManager
-from sqlalchemy_bind_manager._repository import SQLAlchemyAsyncRepository
+from sqlalchemy_bind_manager.repository import SQLAlchemyAsyncRepository
 
-from common.config import AppConfig
+from bootstrap.config import AppConfig
 
 
 class Container(DeclarativeContainer):
@@ -21,7 +21,7 @@ class Container(DeclarativeContainer):
 
     wiring_config = WiringConfiguration(
         packages=[
-            "gateways",
+            "bootstrap",
             "domains",
         ]
     )
