+++ tests/http_app/test_factory.py
@@ -1,19 +1,19 @@
 from unittest.mock import patch
 
-from common.config import AppConfig
+from bootstrap.config import AppConfig
 from http_app import create_app
 
 
 def test_with_default_config() -> None:
     """Test create_app without passing test config."""
-    with patch("common.bootstrap.init_storage", return_value=None):
+    with patch("bootstrap.bootstrap.init_storage", return_value=None):
         app = create_app()
     assert app.debug is False
 
 
 def test_with_debug_config() -> None:
     # We don't need the storage to test the HTTP app
-    with patch("common.bootstrap.init_storage", return_value=None):
+    with patch("bootstrap.bootstrap.init_storage", return_value=None):
         app = create_app(
             test_config=AppConfig(
                 SQLALCHEMY_CONFIG={},
