+++ tests/storage/conftest.py
@@ -1,7 +1,7 @@
 from collections.abc import AsyncIterator
 
 import pytest
-from gateways.storage.SQLAlchemy import init_tables
+from bootstrap.storage.SQLAlchemy import init_tables
 from sqlalchemy.orm import clear_mappers
 from sqlalchemy_bind_manager import SQLAlchemyAsyncConfig, SQLAlchemyBindManager
